fix(web): resolve dart2js compile errors for flutter build web
identity_service.dart: add _toInt64() helper returning dynamic to pass
int values as PlatformInt64 (BigInt on web, int on native) to frb bridge.

home_order_providers.dart: convert PlatformInt64 createdAt/expiresAt to
int before passing to DateTime.fromMillisecondsSinceEpoch via is-BigInt
guard with unnecessary_cast suppression for native builds.

trades_providers.dart: cast else-branch to int in startedAt ternary so
the expression is typed int (not Object) on both platforms.

notifications_provider.dart + core/stubs/path_provider_stub.dart: replace
conditional import target 'dart:html' with a proper stub that provides
getApplicationDocumentsDirectory() so dart2js finds the symbol even though
the kIsWeb branch never calls it at runtime.

+++ lib/features/trades/providers/trades_providers.dart
@@ -106,7 +106,7 @@ TradeListItem _tradeInfoToItem(rust_types.TradeInfo trade) {
   // yet implemented; the BigInt branch guards future correctness.
   final createdAt = trade.startedAt is BigInt
       ? (trade.startedAt as BigInt).toInt()
-      : trade.startedAt;
+      : trade.startedAt as int; // ignore: unnecessary_cast
 
   return TradeListItem(
     orderId: trade.order.id,
